+++ src/main/java/com/alzzaipo/common/jwt/JwtFilter.java
@@ -7,6 +7,9 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
@@ -16,9 +19,6 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-import java.io.IOException;
-import java.util.List;
-
 @Slf4j
 @RequiredArgsConstructor
 public class JwtFilter extends OncePerRequestFilter {
@@ -26,8 +26,9 @@ public class JwtFilter extends OncePerRequestFilter {
     private final JwtUtil jwtUtil;
 
     @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws IOException, ServletException {
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
+        FilterChain filterChain)
+        throws IOException, ServletException {
 
         String authorization = request.getHeader(HttpHeaders.AUTHORIZATION);
 
@@ -42,11 +43,12 @@ public class JwtFilter extends OncePerRequestFilter {
             MemberPrincipal memberPrincipal = createPrincipalFromToken(token);
 
             UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
-                    memberPrincipal,
-                    null,
-                    List.of(new SimpleGrantedAuthority("USER")));
+                memberPrincipal,
+                null,
+                List.of(new SimpleGrantedAuthority("USER")));
 
-            authenticationToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+            authenticationToken.setDetails(
+                new WebAuthenticationDetailsSource().buildDetails(request));
             SecurityContextHolder.getContext().setAuthentication(authenticationToken);
         } catch (ExpiredJwtException e) {
             response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "토큰 만료");
@@ -65,7 +67,25 @@ public class JwtFilter extends OncePerRequestFilter {
 
     private MemberPrincipal createPrincipalFromToken(String token) {
         return new MemberPrincipal(
-                jwtUtil.getMemberUID(token),
-                jwtUtil.getLoginType(token));
+            jwtUtil.getMemberUID(token),
+            jwtUtil.getLoginType(token));
     }
+
+    @Override
+    protected boolean shouldNotFilter(HttpServletRequest request) throws ServletException {
+        List<String> excludePath = Arrays.asList(
+            "/member/verify-account-id",
+            "/member/verify-email",
+            "/member/register",
+            "/member/login",
+            "/ipo/*",
+            "/email/*",
+            "/scraper",
+            "/oauth/kakao/login");
+
+        String path = request.getRequestURI();
+
+        return excludePath.stream().anyMatch(path::startsWith);
+    }
+
 }
